+++ package.json
@@ -14,11 +14,12 @@
     "build": "tsx build.ts",
     "checkts": "tsc --noEmit && tsc --noEmit --project tsconfig.connectors.json",
     "test": "vitest",
-    "lint": "npm run lint:eslint && npm run lint:stylelint && npm run lint:prettier && npm run lint:markdown",
+    "lint": "npm run lint:eslint && npm run lint:stylelint && npm run lint:prettier && npm run lint:markdown && npm run lint:circular",
     "lint:eslint": "eslint .",
     "lint:stylelint": "stylelint \"**/*.scss\"",
     "lint:prettier": "prettier . --check",
     "lint:markdown": "remark .",
+    "lint:circular": "madge --extensions ts,tsx --circular src",
     "prettierfix": "prettier . --write",
     "fixstyle": "stylelint \"**/*.scss\" --fix",
     "translations": "tx pull -m onlytranslated && tsx scripts/translations-prepare.ts"
@@ -52,9 +53,9 @@
     "archiver": "7.0.1",
     "canvas": "3.2.0",
     "chokidar": "4.0.3",
+    "esbuild": "0.25.11",
     "eslint": "9.39.2",
     "eslint-plugin-jsdoc": "59.0.2",
-    "esbuild": "0.25.11",
     "eslint-plugin-react": "7.37.5",
     "eslint-plugin-solid": "0.14.5",
     "eslint-plugin-tsdoc": "0.5.2",
@@ -63,6 +64,7 @@
     "imagemin": "9.0.1",
     "imagemin-jpegtran": "8.0.0",
     "imagemin-pngquant": "10.0.0",
+    "madge": "^8.0.0",
     "prettier": "3.6.2",
     "prettier-config-web-scrobbler": "0.1.0",
     "remark": "15.0.1",
