+++ src/core/background/util.ts
@@ -4,7 +4,6 @@ import browser from 'webextension-polyfill';
 import * as ControllerMode from '@/core/object/controller/controller-mode';
 import * as BrowserStorage from '@/core/storage/browser-storage';
 import { isPrioritizedMode } from '@/core/object/controller/controller';
-import { performUpdateAction } from './action';
 import { sendBackgroundMessage } from '@/util/communication';
 import type { ConnectorMeta } from '../connectors';
 
@@ -108,22 +107,6 @@ export function unlockState(): void {
  */
 export async function setState(data: StateManagement): Promise<void> {
 	return state.setLocking(data);
-}
-
-/**
- * Takes a list of tabs and updates the action state based on the first tab that matches a priority group.
- *
- * @param tabs - tabs from state
- * @param tabId - Currently active tab id if known
- * @returns the tab that was used to update the action state
- */
-export async function updateTabsFromTabList(
-	tabs: ManagerTab[],
-	tabId?: number,
-) {
-	const curTab = await getActiveTabDetails(tabs, tabId);
-	performUpdateAction(curTab);
-	return curTab;
 }
 
 /**
