+++ src/ui/options/components/modal-type.ts
@@ -1,0 +1,10 @@
+/**
+ * Type indicating possible states for modal
+ */
+export type ModalType =
+	| 'savedEdits'
+	| 'regexEdits'
+	| 'cacheEdit'
+	| 'blockedTags'
+	| 'blocklist'
+	| '';
